feat: create new stripe component to handle all stripe transactions only
